+++ aead/src/dev.rs
@@ -8,9 +8,9 @@ macro_rules! new_test {
         #[test]
         fn $name() {
             use aead::{
-                array::{typenum::Unsigned, Array},
-                dev::blobby::Blob6Iterator,
                 Aead, KeyInit, Payload,
+                array::{Array, typenum::Unsigned},
+                dev::blobby::Blob6Iterator,
             };
 
             fn run_test(