+++ elliptic-curve/src/hash2curve/group_digest.rs
@@ -1,6 +1,6 @@
 //! Traits for handling hash to curve.
 
-use super::{hash_to_field, ExpandMsg, FromOkm, MapToCurve};
+use super::{ExpandMsg, FromOkm, MapToCurve, hash_to_field};
 use crate::{CurveArithmetic, ProjectivePoint, Result};
 use group::cofactor::CofactorGroup;