+++ digest/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Mac::new`, `Mac::new_from_slice`, and `Mac::generate_key` methods ([#1173])
 
 [#1173]: https://github.com/RustCrypto/traits/pull/1173
+[#1334]: https://github.com/RustCrypto/traits/pull/1334
 
 ## 0.10.7 (2023-05-19)
 ### Changed